allow partial evaluation of dataframe before prediction
